+++ docs/changes.rst
@@ -10,6 +10,7 @@ Released on ? 2018.
 * Changed repository locking to use third party library.
 * Added support for downloading only strings needing action.
 * Added support for searching in several languages at once.
+* New addon to configure Gettext output wrapping.
 
 weblate 2.19.1
 --------------